Added a Plan class to hold the evaluation plan
